fastapi-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.12.4
pydantic-settings==2.1.0

